@@ -22,13 +22,13 @@ namespace wan24.ObjectValidation
     /// <summary>
     /// Item person TIN validation attribute
     /// </summary>
-    public class ItemPersonTINAttribute : ItemValidationAttribute
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="country">Country</param>
+    /// <param name="target">Validation target</param>
+    public class ItemPersonTINAttribute(Country country = Country.XX, ItemValidationTargets target = ItemValidationTargets.Item)
+        : ItemValidationAttribute(target, new PersonTINAttribute(country))
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="country">Country</param>
-        /// <param name="target">Validation target</param>
-        public ItemPersonTINAttribute(Country country = Country.XX, ItemValidationTargets target = ItemValidationTargets.Item) : base(target, new PersonTINAttribute(country)) { }
     }
 }
