@@ -22,13 +22,13 @@ namespace wan24.ObjectValidation
     /// <summary>
     /// Item SSN validation attribute
     /// </summary>
-    public class ItemSSNAttribute : ItemValidationAttribute
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="country">Country</param>
+    /// <param name="target">Validation target</param>
+    public class ItemSSNAttribute(Country country = Country.XX, ItemValidationTargets target = ItemValidationTargets.Item)
+        : ItemValidationAttribute(target, new SSNAttribute(country))
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="country">Country</param>
-        /// <param name="target">Validation target</param>
-        public ItemSSNAttribute(Country country = Country.XX, ItemValidationTargets target = ItemValidationTargets.Item):base(target, new SSNAttribute(country)) { }
     }
 }
