@@ -3,15 +3,13 @@
     /// <summary>
     /// ABA RTN item validation attribute
     /// </summary>
-    public class ItemAbaRtnAttribute : ItemValidationAttribute
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="normalize">Normalize?</param>
+    /// <param name="target">Validation target</param>
+    public class ItemAbaRtnAttribute(bool normalize = true, ItemValidationTargets target = ItemValidationTargets.Item) : ItemValidationAttribute(target, new AbaRtnAttribute(normalize))
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="normalize">Normalize?</param>
-        /// <param name="target">Validation target</param>
-        public ItemAbaRtnAttribute(bool normalize = true, ItemValidationTargets target = ItemValidationTargets.Item) : base(target, new AbaRtnAttribute(normalize)) { }
-
         /// <summary>
         /// Supported ABA RTN formats
         /// </summary>
