@@ -3,13 +3,12 @@
     /// <summary>
     /// Item amount validation attribute
     /// </summary>
-    public class ItemAmountAttribute : ItemValidationAttribute
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="currency">Currency</param>
+    /// <param name="target">Target validation</param>
+    public class ItemAmountAttribute(string currency, ItemValidationTargets target = ItemValidationTargets.Item) : ItemValidationAttribute(target, new AmountAttribute(currency))
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="currency">Currency</param>
-        /// <param name="target">Target validation</param>
-        public ItemAmountAttribute(string currency, ItemValidationTargets target = ItemValidationTargets.Item) : base(target, new AmountAttribute(currency)) { }
     }
 }
