@@ -5,12 +5,11 @@ namespace wan24.ObjectValidation
     /// <summary>
     /// Item data type validation attribute
     /// </summary>
-    public class ItemDataTypeAttribute : ItemValidationAttribute
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="dataType">Data type</param>
+    public class ItemDataTypeAttribute(DataType dataType) : ItemValidationAttribute(new DataTypeAttribute(dataType))
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="dataType">Data type</param>
-        public ItemDataTypeAttribute(DataType dataType) : base(new DataTypeAttribute(dataType)) { }
     }
 }
