@@ -5,12 +5,11 @@ namespace wan24.ObjectValidation
     /// <summary>
     /// Item email address attribute
     /// </summary>
-    public class ItemEmailAddressAttribute : ItemValidationAttribute
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="target">Validation target</param>
+    public class ItemEmailAddressAttribute(ItemValidationTargets target = ItemValidationTargets.Item) : ItemValidationAttribute(target, new EmailAddressAttribute())
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="target">Validation target</param>
-        public ItemEmailAddressAttribute(ItemValidationTargets target = ItemValidationTargets.Item) : base(target, new EmailAddressAttribute()) { }
     }
 }
