@@ -5,13 +5,13 @@ namespace wan24.ObjectValidation
     /// <summary>
     /// Item IP address validation attribute
     /// </summary>
-    public class ItemIpAttribute : ItemValidationAttribute
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="family">Address family</param>
+    /// <param name="target">Validation target</param>
+    public class ItemIpAttribute(AddressFamily family = AddressFamily.InterNetwork, ItemValidationTargets target = ItemValidationTargets.Item)
+        : ItemValidationAttribute(target,new IpAttribute(family))
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="family">Address family</param>
-        /// <param name="target">Validation target</param>
-        public ItemIpAttribute(AddressFamily family = AddressFamily.InterNetwork, ItemValidationTargets target = ItemValidationTargets.Item):base(target,new IpAttribute(family)) { }
     }
 }
