@@ -5,13 +5,12 @@ namespace wan24.ObjectValidation
     /// <summary>
     /// Maximum item length attribute
     /// </summary>
-    public class ItemMaxLengthAttribute : ItemValidationAttribute
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="len">Length</param>
+    /// <param name="target">Validation target</param>
+    public class ItemMaxLengthAttribute(int len, ItemValidationTargets target = ItemValidationTargets.Item) : ItemValidationAttribute(target, new MaxLengthAttribute(len))
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="len">Length</param>
-        /// <param name="target">Validation target</param>
-        public ItemMaxLengthAttribute(int len, ItemValidationTargets target = ItemValidationTargets.Item) : base(target, new MaxLengthAttribute(len)) { }
     }
 }
