@@ -5,13 +5,12 @@ namespace wan24.ObjectValidation
     /// <summary>
     /// Minimum item length attribute
     /// </summary>
-    public class ItemMinLengthAttribute : ItemValidationAttribute
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="len">Length</param>
+    /// <param name="target">Validation target</param>
+    public class ItemMinLengthAttribute(int len, ItemValidationTargets target = ItemValidationTargets.Item) : ItemValidationAttribute(target, new MinLengthAttribute(len))
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="len">Length</param>
-        /// <param name="target">Validation target</param>
-        public ItemMinLengthAttribute(int len, ItemValidationTargets target = ItemValidationTargets.Item) : base(target, new MinLengthAttribute(len)) { }
     }
 }
