@@ -5,12 +5,11 @@ namespace wan24.ObjectValidation
     /// <summary>
     /// Item phone number attribute
     /// </summary>
-    public class ItemPhoneAttribute : ItemValidationAttribute
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="target">Validation target</param>
+    public class ItemPhoneAttribute(ItemValidationTargets target = ItemValidationTargets.Item) : ItemValidationAttribute(target, new PhoneAttribute())
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="target">Validation target</param>
-        public ItemPhoneAttribute(ItemValidationTargets target = ItemValidationTargets.Item) : base(target, new PhoneAttribute()) { }
     }
 }
