@@ -5,13 +5,13 @@ namespace wan24.ObjectValidation
     /// <summary>
     /// Item regular expression attribute
     /// </summary>
-    public class ItemRegularExpressionAttribute : ItemValidationAttribute
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="pattern">Pattern</param>
+    /// <param name="target">Validation target</param>
+    public class ItemRegularExpressionAttribute(string pattern, ItemValidationTargets target = ItemValidationTargets.Item)
+        : ItemValidationAttribute(target, new RegularExpressionAttribute(pattern))
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="pattern">Pattern</param>
-        /// <param name="target">Validation target</param>
-        public ItemRegularExpressionAttribute(string pattern, ItemValidationTargets target = ItemValidationTargets.Item) : base(target, new RegularExpressionAttribute(pattern)) { }
     }
 }
