@@ -5,12 +5,11 @@ namespace wan24.ObjectValidation
     /// <summary>
     /// Item required attribute
     /// </summary>
-    public class ItemRequiredAttribute : ItemValidationAttribute
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="target">Validation target</param>
+    public class ItemRequiredAttribute(ItemValidationTargets target = ItemValidationTargets.Item) : ItemValidationAttribute(target, new RequiredAttribute())
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="target">Validation target</param>
-        public ItemRequiredAttribute(ItemValidationTargets target = ItemValidationTargets.Item) : base(target, new RequiredAttribute()) { }
     }
 }
