@@ -5,12 +5,11 @@ namespace wan24.ObjectValidation
     /// <summary>
     /// Item URL attribute
     /// </summary>
-    public class ItemUrlAttribute : ItemValidationAttribute
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="target">Validation target</param>
+    public class ItemUrlAttribute(ItemValidationTargets target = ItemValidationTargets.Item) : ItemValidationAttribute(target, new UrlAttribute())
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="target">Validation target</param>
-        public ItemUrlAttribute(ItemValidationTargets target = ItemValidationTargets.Item) : base(target, new UrlAttribute()) { }
     }
 }
