@@ -5,22 +5,16 @@ namespace wan24.ObjectValidation
     /// <summary>
     /// Item template validation attribute
     /// </summary>
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="template">Template key</param>
+    /// <param name="target">Validation target</param>
     [AttributeUsage(AttributeTargets.Property, AllowMultiple = true, Inherited = true)]
-    public class ItemValidationTemplateAttribute : Attribute, IItemValidationAttribute
+    public class ItemValidationTemplateAttribute(string template, ItemValidationTargets target = ItemValidationTargets.Item) : Attribute(), IItemValidationAttribute
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="template">Template key</param>
-        /// <param name="target">Validation target</param>
-        public ItemValidationTemplateAttribute(string template, ItemValidationTargets target = ItemValidationTargets.Item) : base()
-        {
-            ValidationTarget = target;
-            Template = template;
-        }
-
         /// <inheritdoc/>
-        public ItemValidationTargets ValidationTarget { get; }
+        public ItemValidationTargets ValidationTarget { get; } = target;
 
         /// <inheritdoc/>
         public int ArrayLevel { get; set; }
@@ -28,7 +22,7 @@ namespace wan24.ObjectValidation
         /// <summary>
         /// Template key
         /// </summary>
-        public string Template { get; }
+        public string Template { get; } = template;
 
         /// <inheritdoc/>
         public virtual ValidationResult? GetValidationResult(object? value, ValidationContext validationContext, IServiceProvider? serviceProvider)
@@ -37,7 +31,7 @@ namespace wan24.ObjectValidation
         /// <inheritdoc/>
         public virtual IEnumerable<ValidationResult> MultiValidation(object? value, ValidationContext validationContext, IServiceProvider? serviceProvider = null)
         {
-            if (!ValidationTemplates.ItemValidations.TryGetValue(Template, out List<IItemValidationAttribute>? attrs))
+            if (!ValidationTemplates.ItemValidations.TryGetValue(Template, out List<IItemValidationAttribute>? attributes))
             {
                 yield return new(
                     validationContext.MemberName is null ? $"Validation template \"{Template}\" not found" : $"{validationContext.MemberName}: Validation template \"{Template}\" not found",
@@ -46,7 +40,7 @@ namespace wan24.ObjectValidation
             }
             else
             {
-                foreach (IItemValidationAttribute attr in from a in attrs
+                foreach (IItemValidationAttribute attr in from a in attributes
                                                           where a.ValidationTarget == ValidationTarget &&
                                                             a.ArrayLevel == ArrayLevel
                                                           select a)
