@@ -3,12 +3,11 @@
     /// <summary>
     /// Item XML validation attribute
     /// </summary>
-    public class ItemXmlAttribute : ItemValidationAttribute
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="target">Validation target</param>
+    public class ItemXmlAttribute(ItemValidationTargets target = ItemValidationTargets.Item) : ItemValidationAttribute(target, new XmlAttribute())
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="target">Validation target</param>
-        public ItemXmlAttribute(ItemValidationTargets target = ItemValidationTargets.Item) : base(target, new XmlAttribute()) { }
     }
 }
