@@ -5,23 +5,21 @@ namespace wan24.ObjectValidation
     /// <summary>
     /// Template validation attribute
     /// </summary>
-    public class ValidationTemplateAttribute : ValidationAttribute, IMultipleValidations
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="template">Template key</param>
+    public class ValidationTemplateAttribute(string template) : ValidationAttribute(), IMultipleValidations
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="template">Template key</param>
-        public ValidationTemplateAttribute(string template) : base() => Template = template;
-
         /// <summary>
         /// Template key
         /// </summary>
-        public string Template { get; }
+        public string Template { get; } = template;
 
         /// <inheritdoc/>
         public virtual IEnumerable<ValidationResult> MultiValidation(object? value, ValidationContext validationContext, IServiceProvider? serviceProvider = null)
         {
-            if (!ValidationTemplates.PropertyValidations.TryGetValue(Template, out List<ValidationAttribute>? attrs))
+            if (!ValidationTemplates.PropertyValidations.TryGetValue(Template, out List<ValidationAttribute>? attributes))
             {
                 yield return new(
                     ErrorMessage ?? (validationContext.MemberName is null ? $"Validation template \"{Template}\" not found" : $"{validationContext.MemberName}: Validation template \"{Template}\" not found"),
@@ -30,7 +28,7 @@ namespace wan24.ObjectValidation
             }
             else
             {
-                foreach (ValidationAttribute attr in attrs)
+                foreach (ValidationAttribute attr in attributes)
                     if (attr.GetValidationResult(value, validationContext) is ValidationResult result)
                         yield return result;
             }
