@@ -10,11 +10,11 @@ namespace wan24.ObjectValidation
         /// <summary>
         /// Property validation templates (key is the template name, value a list of validation attributes to apply)
         /// </summary>
-        public static readonly Dictionary<string, List<ValidationAttribute>> PropertyValidations = new();
+        public static readonly Dictionary<string, List<ValidationAttribute>> PropertyValidations = [];
 
         /// <summary>
         /// Item validation templates (key is the template name, value a list of validation attributes to apply)
         /// </summary>
-        public static readonly Dictionary<string, List<IItemValidationAttribute>> ItemValidations = new();
+        public static readonly Dictionary<string, List<IItemValidationAttribute>> ItemValidations = [];
     }
 }
